Use old name for Python nodes
So existing workflows will keep working.
