+++ plugin/META-INF/MANIFEST.MF
@@ -21,5 +21,5 @@ Bundle-ClassPath: lib/gson-2.6.2.jar,
  lib/swagger-annotations-1.5.12.jar,
  .
 Export-Package: nl.esciencecenter.e3dchem.kripodb,
- nl.esciencecenter.e3dchem.kripodb.py.fragments,
- nl.esciencecenter.e3dchem.kripodb.py.fragmentsbysimilarity
+ nl.esciencecenter.e3dchem.kripodb.fragments,
+ nl.esciencecenter.e3dchem.kripodb.fragmentsbysimilarity
