+++ plugin/src/java/nl/esciencecenter/e3dchem/kripodb/fragments/FragmentByIdDialog.java
@@ -1,4 +1,4 @@
-package nl.esciencecenter.e3dchem.kripodb.py.fragments;
+package nl.esciencecenter.e3dchem.kripodb.fragments;
 
 import java.util.Set;
