+++ plugin/src/java/nl/esciencecenter/e3dchem/kripodb/fragments/FragmentByIdFactory.java
@@ -1,4 +1,4 @@
-package nl.esciencecenter.e3dchem.kripodb.py.fragmentsbysimilarity;
+package nl.esciencecenter.e3dchem.kripodb.fragments;
 
 import org.knime.core.node.NodeDialogPane;
 import org.knime.core.node.NodeFactory;
@@ -8,14 +8,14 @@ import org.knime.core.node.NodeView;
  * <code>NodeFactory</code> for the "FragmentBySimilarity" Node.
  *
  */
-public class FragmentBySimilarityFactory extends NodeFactory<FragmentBySimilarityModel> {
+public class FragmentByIdFactory extends NodeFactory<FragmentByIdModel> {
 
 	/**
 	 * {@inheritDoc}
 	 */
 	@Override
-	public FragmentBySimilarityModel createNodeModel() {
-		return new FragmentBySimilarityModel();
+	public FragmentByIdModel createNodeModel() {
+		return new FragmentByIdModel();
 	}
 
 	/**
@@ -39,14 +39,11 @@ public class FragmentBySimilarityFactory extends NodeFactory<FragmentBySimilarit
 	 */
 	@Override
 	public NodeDialogPane createNodeDialogPane() {
-		return new FragmentBySimilarityDialog();
+		return new FragmentByIdDialog();
 	}
 
-	/**
-	 * {@inheritDoc}
-	 */
 	@Override
-	public NodeView<FragmentBySimilarityModel> createNodeView(int viewIndex, FragmentBySimilarityModel nodeModel) {
+	public NodeView<FragmentByIdModel> createNodeView(int viewIndex, FragmentByIdModel nodeModel) {
 		return null;
 	}
